refactor: List<Product>를 일급 클래스로 생성, 로직 분리
