+++ src/main/java/vendingmachine/firstclass/Products.java
@@ -1,0 +1,37 @@
+package vendingmachine.firstclass;
+
+import vendingmachine.model.Product;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import static vendingmachine.message.Error.NOT_ENOUGH_TO_BUY;
+import static vendingmachine.message.Error.NO_SUCH_PRODUCT_EXIST;
+
+public class Products {
+
+	private List<Product> products = new ArrayList<>();
+
+	public Products(List<Product> products) {
+		this.products = products;
+	}
+
+	public int sellProduct(String name, int money) {
+		Product product = getProductByName(name);
+		return countOutProduct(product, money);
+	}
+
+	private Product getProductByName(String name) {
+		return products.stream()
+				.filter(product -> product.getName().equals(name))
+				.findAny()
+				.orElseThrow(() -> new IllegalArgumentException(NO_SUCH_PRODUCT_EXIST));
+	}
+
+	private int countOutProduct(Product product, int money) {
+		if (!product.enoughMoneyToBuy(money)) {
+			throw new IllegalArgumentException(NOT_ENOUGH_TO_BUY);
+		}
+		return product.purchaseOne(money);
+	}
+}
